+++ src/core/index.ts
@@ -1,3 +1,3 @@
-export * from './jwt'
+export { default as JWTManager } from './jwt'
 export * from './permission'
 export * from './role'
