--- src/core/jwt/jwt.model.ts
@@ -1,3 +1,0 @@
-import type { User } from '@prisma/client'
-
-export type JWTUserModel = Pick<User, 'id' | 'username'>
