+++ src/middlewares/index.ts
@@ -1,4 +1,4 @@
 export * from './auth'
+export * from './error'
 export * from './i18n'
 export * from './logger'
-export * from './routes'
