--- src/middlewares/routes/index.ts
@@ -1,2 +1,0 @@
-export * from './404'
-export * from './500'
