--- src/shared/config/global-config.type.ts
@@ -1,1 +1,0 @@
-export type ENV = 'development' | 'production' | 'test'
