--- src/shared/file-storage/extension.private.ts
@@ -1,17 +1,0 @@
-export const extensionMap = new Map([
-  ['text/plain', '.txt'],
-  ['image/jpeg', '.jpg'],
-  ['image/png', '.png'],
-  ['image/gif', '.gif'],
-  ['image/bmp', '.bmp'],
-  ['image/webp', '.webp'],
-  ['image/svg+xml', '.svg'],
-  ['application/pdf', '.pdf'],
-  ['application/msword', '.doc'],
-  ['application/vnd.openxmlformats-officedocument.wordprocessingml.document', '.docx'],
-  ['application/vnd.ms-excel', '.xls'],
-  ['application/vnd.openxmlformats-officedocument.spreadsheetml.sheet', '.xlsx'],
-  ['application/vnd.ms-powerpoint', '.ppt'],
-  ['application/vnd.openxmlformats-officedocument.presentationml.presentation', '.pptx'],
-  ['application/zip', '.zip']
-])
