--- src/shared/file-storage/index.ts
@@ -1,1 +1,0 @@
-export * from './file-storage'
