+++ src/shared/index.ts
@@ -1,10 +1,11 @@
 export * from './config'
 export * from './enum'
-export * from './file-storage'
+export * from './env'
 export * from './i18n'
 export * from './logger'
 export * from './password'
 export * from './prisma'
 export * from './random'
+export * from './storage'
 export * from './time'
 export * from './uuid'
