--- src/shared/logger/chalk.private.ts
@@ -1,4 +1,0 @@
-import chalk from 'chalk'
-
-export const chalkPrimary = chalk.green
-export const chalkError = chalk.red
