--- src/shared/logger/index.ts
@@ -1,2 +1,0 @@
-export * from './log'
-export * from './log.constants'
