--- src/shared/logger/log.constants.ts
@@ -1,4 +1,0 @@
-import { GlobalAppConfig } from '../config'
-import { getCurrentTime } from '../time'
-
-export const LOG_PREFIX = `[${GlobalAppConfig.APP_NAME} - ${getCurrentTime('HH:mm:ss')}]`
