+++ src/shared/time.ts
@@ -1,5 +1,8 @@
 import dayjs from 'dayjs'
 
-import type { TimeFormatter } from './time.type'
+import type { TimeFormatter } from '@/types'
 
+/**
+ * 获取带格式化的当前时间
+ */
 export const getCurrentTime = (formatter: TimeFormatter = 'YYYY-MM-DD HH:mm:ss') => dayjs(Date.now()).format(formatter)
