+++ src/types/jwt.ts
@@ -1,0 +1,3 @@
+import type { User } from '@prisma/client'
+
+export type JWTModel = Pick<User, 'id' | 'username'>
